[DF] Move RLoopManager to its own header and source files
