+++ tree/dataframe/src/RLoopManager.cxx
@@ -1,0 +1,425 @@
+#include "RConfigure.h" // R__USE_IMT
+#include "ROOT/RDFNodes.hxx"
+#include "ROOT/RLoopManager.hxx"
+#include "ROOT/RSlotStack.hxx"
+#include "ROOT/TTreeProcessorMT.hxx"
+#include "RtypesCore.h" // Long64_t
+#include "TError.h"
+#include "TInterpreter.h"
+#include "TROOT.h" // IsImplicitMTEnabled
+#include "TTreeReader.h"
+
+#ifdef R__USE_IMT
+#include "ROOT/TThreadExecutor.hxx"
+#endif
+
+#include <functional>
+#include <memory>
+#include <stdexcept>
+#include <string>
+#include <vector>
+
+using namespace ROOT::Detail::RDF;
+using namespace ROOT::Internal::RDF;
+
+RLoopManager::RLoopManager(TTree *tree, const ColumnNames_t &defaultBranches)
+   : fTree(std::shared_ptr<TTree>(tree, [](TTree *) {})), fDefaultColumns(defaultBranches),
+     fNSlots(RDFInternal::GetNSlots()),
+     fLoopType(ROOT::IsImplicitMTEnabled() ? ELoopType::kROOTFilesMT : ELoopType::kROOTFiles)
+{
+}
+
+RLoopManager::RLoopManager(ULong64_t nEmptyEntries)
+   : fNEmptyEntries(nEmptyEntries), fNSlots(RDFInternal::GetNSlots()),
+     fLoopType(ROOT::IsImplicitMTEnabled() ? ELoopType::kNoFilesMT : ELoopType::kNoFiles)
+{
+}
+
+RLoopManager::RLoopManager(std::unique_ptr<RDataSource> ds, const ColumnNames_t &defaultBranches)
+   : fDefaultColumns(defaultBranches), fNSlots(RDFInternal::GetNSlots()),
+     fLoopType(ROOT::IsImplicitMTEnabled() ? ELoopType::kDataSourceMT : ELoopType::kDataSource),
+     fDataSource(std::move(ds))
+{
+   fDataSource->SetNSlots(fNSlots);
+}
+
+/// Run event loop with no source files, in parallel.
+void RLoopManager::RunEmptySourceMT()
+{
+#ifdef R__USE_IMT
+   RSlotStack slotStack(fNSlots);
+   // Working with an empty tree.
+   // Evenly partition the entries according to fNSlots. Produce around 2 tasks per slot.
+   const auto nEntriesPerSlot = fNEmptyEntries / (fNSlots * 2);
+   auto remainder = fNEmptyEntries % (fNSlots * 2);
+   std::vector<std::pair<ULong64_t, ULong64_t>> entryRanges;
+   ULong64_t start = 0;
+   while (start < fNEmptyEntries) {
+      ULong64_t end = start + nEntriesPerSlot;
+      if (remainder > 0) {
+         ++end;
+         --remainder;
+      }
+      entryRanges.emplace_back(start, end);
+      start = end;
+   }
+
+   // Each task will generate a subrange of entries
+   auto genFunction = [this, &slotStack](const std::pair<ULong64_t, ULong64_t> &range) {
+      auto slot = slotStack.GetSlot();
+      InitNodeSlots(nullptr, slot);
+      for (auto currEntry = range.first; currEntry < range.second; ++currEntry) {
+         RunAndCheckFilters(slot, currEntry);
+      }
+      CleanUpTask(slot);
+      slotStack.ReturnSlot(slot);
+   };
+
+   ROOT::TThreadExecutor pool;
+   pool.Foreach(genFunction, entryRanges);
+
+#endif // not implemented otherwise
+}
+
+/// Run event loop with no source files, in sequence.
+void RLoopManager::RunEmptySource()
+{
+   InitNodeSlots(nullptr, 0);
+   for (ULong64_t currEntry = 0; currEntry < fNEmptyEntries && fNStopsReceived < fNChildren; ++currEntry) {
+      RunAndCheckFilters(0, currEntry);
+   }
+}
+
+/// Run event loop over one or multiple ROOT files, in parallel.
+void RLoopManager::RunTreeProcessorMT()
+{
+#ifdef R__USE_IMT
+   RSlotStack slotStack(fNSlots);
+   auto tp = std::make_unique<ROOT::TTreeProcessorMT>(*fTree);
+
+   tp->Process([this, &slotStack](TTreeReader &r) -> void {
+      auto slot = slotStack.GetSlot();
+      InitNodeSlots(&r, slot);
+      // recursive call to check filters and conditionally execute actions
+      while (r.Next()) {
+         RunAndCheckFilters(slot, r.GetCurrentEntry());
+      }
+      CleanUpTask(slot);
+      slotStack.ReturnSlot(slot);
+   });
+#endif // no-op otherwise (will not be called)
+}
+
+/// Run event loop over one or multiple ROOT files, in sequence.
+void RLoopManager::RunTreeReader()
+{
+   TTreeReader r(fTree.get());
+   if (0 == fTree->GetEntriesFast())
+      return;
+   InitNodeSlots(&r, 0);
+
+   // recursive call to check filters and conditionally execute actions
+   // in the non-MT case processing can be stopped early by ranges, hence the check on fNStopsReceived
+   while (r.Next() && fNStopsReceived < fNChildren) {
+      RunAndCheckFilters(0, r.GetCurrentEntry());
+   }
+   fTree->GetEntry(0);
+}
+
+/// Run event loop over data accessed through a DataSource, in sequence.
+void RLoopManager::RunDataSource()
+{
+   R__ASSERT(fDataSource != nullptr);
+   fDataSource->Initialise();
+   auto ranges = fDataSource->GetEntryRanges();
+   while (!ranges.empty()) {
+      InitNodeSlots(nullptr, 0u);
+      fDataSource->InitSlot(0u, 0ull);
+      for (const auto &range : ranges) {
+         auto end = range.second;
+         for (auto entry = range.first; entry < end; ++entry) {
+            if (fDataSource->SetEntry(0u, entry)) {
+               RunAndCheckFilters(0u, entry);
+            }
+         }
+      }
+      fDataSource->FinaliseSlot(0u);
+      ranges = fDataSource->GetEntryRanges();
+   }
+   fDataSource->Finalise();
+}
+
+/// Run event loop over data accessed through a DataSource, in parallel.
+void RLoopManager::RunDataSourceMT()
+{
+#ifdef R__USE_IMT
+   R__ASSERT(fDataSource != nullptr);
+   RSlotStack slotStack(fNSlots);
+   ROOT::TThreadExecutor pool;
+
+   // Each task works on a subrange of entries
+   auto runOnRange = [this, &slotStack](const std::pair<ULong64_t, ULong64_t> &range) {
+      const auto slot = slotStack.GetSlot();
+      InitNodeSlots(nullptr, slot);
+      fDataSource->InitSlot(slot, range.first);
+      const auto end = range.second;
+      for (auto entry = range.first; entry < end; ++entry) {
+         if (fDataSource->SetEntry(slot, entry)) {
+            RunAndCheckFilters(slot, entry);
+         }
+      }
+      CleanUpTask(slot);
+      fDataSource->FinaliseSlot(slot);
+      slotStack.ReturnSlot(slot);
+   };
+
+   fDataSource->Initialise();
+   auto ranges = fDataSource->GetEntryRanges();
+   while (!ranges.empty()) {
+      pool.Foreach(runOnRange, ranges);
+      ranges = fDataSource->GetEntryRanges();
+   }
+   fDataSource->Finalise();
+#endif // not implemented otherwise (never called)
+}
+
+/// Execute actions and make sure named filters are called for each event.
+/// Named filters must be called even if the analysis logic would not require it, lest they report confusing results.
+void RLoopManager::RunAndCheckFilters(unsigned int slot, Long64_t entry)
+{
+   for (auto &actionPtr : fBookedActions)
+      actionPtr->Run(slot, entry);
+   for (auto &namedFilterPtr : fBookedNamedFilters)
+      namedFilterPtr->CheckFilters(slot, entry);
+   for (auto &callback : fCallbacks)
+      callback(slot);
+}
+
+/// Build TTreeReaderValues for all nodes
+/// This method loops over all filters, actions and other booked objects and
+/// calls their `InitRDFValues` methods. It is called once per node per slot, before
+/// running the event loop. It also informs each node of the TTreeReader that
+/// a particular slot will be using.
+void RLoopManager::InitNodeSlots(TTreeReader *r, unsigned int slot)
+{
+   for (auto &ptr : fBookedActions)
+      ptr->InitSlot(r, slot);
+   for (auto &ptr : fBookedFilters)
+      ptr->InitSlot(r, slot);
+   for (auto &callback : fCallbacksOnce)
+      callback(slot);
+}
+
+/// Initialize all nodes of the functional graph before running the event loop.
+/// This method is called once per event-loop and performs generic initialization
+/// operations that do not depend on the specific processing slot (i.e. operations
+/// that are common for all threads).
+void RLoopManager::InitNodes()
+{
+   EvalChildrenCounts();
+   for (auto column : fCustomColumns)
+      column->InitNode();
+   for (auto &filter : fBookedFilters)
+      filter->InitNode();
+   for (auto &range : fBookedRanges)
+      range->InitNode();
+   for (auto &ptr : fBookedActions)
+      ptr->Initialize();
+}
+
+/// Perform clean-up operations. To be called at the end of each event loop.
+void RLoopManager::CleanUpNodes()
+{
+   fMustRunNamedFilters = false;
+
+   // forget RActions and detach TResultProxies
+   for (auto &ptr : fBookedActions)
+      ptr->Finalize();
+
+   fRunActions.insert(fRunActions.begin(), fBookedActions.begin(), fBookedActions.end());
+   fBookedActions.clear();
+
+   // reset children counts
+   fNChildren = 0;
+   fNStopsReceived = 0;
+   for (auto &ptr : fBookedFilters)
+      ptr->ResetChildrenCount();
+   for (auto &ptr : fBookedRanges)
+      ptr->ResetChildrenCount();
+
+   fCallbacks.clear();
+   fCallbacksOnce.clear();
+}
+
+/// Perform clean-up operations. To be called at the end of each task execution.
+void RLoopManager::CleanUpTask(unsigned int slot)
+{
+   for (auto &ptr : fBookedActions)
+      ptr->FinalizeSlot(slot);
+   for (auto &ptr : fBookedFilters)
+      ptr->ClearTask(slot);
+}
+
+/// Jit all actions that required runtime column type inference, and clean the `fToJit` member variable.
+void RLoopManager::BuildJittedNodes()
+{
+   auto error = TInterpreter::EErrorCode::kNoError;
+   gInterpreter->Calc(fToJit.c_str(), &error);
+   if (TInterpreter::EErrorCode::kNoError != error) {
+      std::string exceptionText =
+         "An error occurred while jitting. The lines above might indicate the cause of the crash\n";
+      throw std::runtime_error(exceptionText.c_str());
+   }
+   fToJit.clear();
+}
+
+/// Trigger counting of number of children nodes for each node of the functional graph.
+/// This is done once before starting the event loop. Each action sends an `increase children count` signal
+/// upstream, which is propagated until RLoopManager. Each time a node receives the signal, in increments its
+/// children counter. Each node only propagates the signal once, even if it receives it multiple times.
+/// Named filters also send an `increase children count` signal, just like actions, as they always execute during
+/// the event loop so the graph branch they belong to must count as active even if it does not end in an action.
+void RLoopManager::EvalChildrenCounts()
+{
+   for (auto &actionPtr : fBookedActions)
+      actionPtr->TriggerChildrenCount();
+   for (auto &namedFilterPtr : fBookedNamedFilters)
+      namedFilterPtr->TriggerChildrenCount();
+}
+
+unsigned int RLoopManager::GetNextID() const
+{
+   static unsigned int id = 0;
+   ++id;
+   return id;
+}
+
+/// Start the event loop with a different mechanism depending on IMT/no IMT, data source/no data source.
+/// Also perform a few setup and clean-up operations (jit actions if necessary, clear booked actions after the loop...).
+void RLoopManager::Run()
+{
+   if (!fToJit.empty())
+      BuildJittedNodes();
+
+   InitNodes();
+
+   switch (fLoopType) {
+   case ELoopType::kNoFilesMT: RunEmptySourceMT(); break;
+   case ELoopType::kROOTFilesMT: RunTreeProcessorMT(); break;
+   case ELoopType::kDataSourceMT: RunDataSourceMT(); break;
+   case ELoopType::kNoFiles: RunEmptySource(); break;
+   case ELoopType::kROOTFiles: RunTreeReader(); break;
+   case ELoopType::kDataSource: RunDataSource(); break;
+   }
+
+   CleanUpNodes();
+}
+
+/// Return the list of default columns -- empty if none was provided when constructing the RDataFrame
+const ColumnNames_t &RLoopManager::GetDefaultColumnNames() const
+{
+   return fDefaultColumns;
+}
+
+TTree *RLoopManager::GetTree() const
+{
+   return fTree.get();
+}
+
+void RLoopManager::Book(RDFInternal::RActionBase *actionPtr)
+{
+   fBookedActions.emplace_back(actionPtr);
+}
+
+void RLoopManager::Deregister(RDFInternal::RActionBase *actionPtr)
+{
+   RDFInternal::Erase(actionPtr, fRunActions);
+   RDFInternal::Erase(actionPtr, fBookedActions);
+}
+
+void RLoopManager::Book(RFilterBase *filterPtr)
+{
+   fBookedFilters.emplace_back(filterPtr);
+   if (filterPtr->HasName()) {
+      fBookedNamedFilters.emplace_back(filterPtr);
+      fMustRunNamedFilters = true;
+   }
+}
+
+void RLoopManager::Deregister(RFilterBase *filterPtr)
+{
+   RDFInternal::Erase(filterPtr, fBookedFilters);
+   RDFInternal::Erase(filterPtr, fBookedNamedFilters);
+}
+
+void RLoopManager::Book(RRangeBase *rangePtr)
+{
+   fBookedRanges.emplace_back(rangePtr);
+}
+
+void RLoopManager::Deregister(RRangeBase *rangePtr)
+{
+   RDFInternal::Erase(rangePtr, fBookedRanges);
+}
+
+// dummy call, end of recursive chain of calls
+bool RLoopManager::CheckFilters(unsigned int, Long64_t)
+{
+   return true;
+}
+
+/// Call `FillReport` on all booked filters
+void RLoopManager::Report(ROOT::RDF::RCutFlowReport &rep) const
+{
+   for (const auto &fPtr : fBookedNamedFilters)
+      fPtr->FillReport(rep);
+}
+
+void RLoopManager::RegisterCallback(ULong64_t everyNEvents, std::function<void(unsigned int)> &&f)
+{
+   if (everyNEvents == 0ull)
+      fCallbacksOnce.emplace_back(std::move(f), fNSlots);
+   else
+      fCallbacks.emplace_back(everyNEvents, std::move(f), fNSlots);
+}
+
+std::vector<std::string> RLoopManager::GetFiltersNames()
+{
+   std::vector<std::string> filters;
+   for (auto &filter : fBookedFilters) {
+      auto name = (filter->HasName() ? filter->GetName() : "Unnamed Filter");
+      filters.push_back(name);
+   }
+   return filters;
+}
+
+std::vector<RDFInternal::RActionBase *> RLoopManager::GetAllActions(){
+   std::vector<RDFInternal::RActionBase *> actions;
+   actions.insert(actions.begin(), fBookedActions.begin(), fBookedActions.end());
+   actions.insert(actions.begin(), fRunActions.begin(), fRunActions.end());
+   return actions;
+}
+
+RRangeBase::RRangeBase(RLoopManager *implPtr, unsigned int start, unsigned int stop, unsigned int stride,
+                       const unsigned int nSlots)
+   : RNodeBase(implPtr), fStart(start), fStop(stop), fStride(stride), fNSlots(nSlots)
+{
+}
+
+std::shared_ptr<ROOT::Internal::RDF::GraphDrawing::GraphNode> RLoopManager::GetGraph()
+{
+   std::string name;
+   if(fDataSource){
+      name= fDataSource->GetDataSourceType();
+   }else if (fTree){
+      name = fTree->GetName();
+   }else{
+      name = std::to_string(fNEmptyEntries);
+   }
+
+   auto thisNode = std::make_shared<ROOT::Internal::RDF::GraphDrawing::GraphNode>(name);
+   thisNode->SetRoot();
+   thisNode->SetCounter(0);
+   return thisNode;
+}
+
